+++ packages/tool/site/client/src/components/prescription-tracker/prescriptionSearchForm.tsx
@@ -3,7 +3,7 @@ import {Field, Formik, FormikErrors} from "formik"
 import {Button, Form, Label} from "nhsuk-react-components"
 import {MaskedInput} from "nhsuk-react-components-extensions"
 import ButtonList from "../common/buttonList"
-import {PrescriptionSearchCriteria} from "../../common/prescription"
+import {PrescriptionSearchCriteria} from "../../common/prescription-search"
 import {BackButton} from "../common/backButton"
 import SelectField, {convertCodingsToOptions} from "../common/selectField"
 import {VALUE_SET_PRESCRIPTION_STATUS} from "../../fhir/reference-data/valueSets"
